requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
msal==1.31.1
requests==2.32.3
pydantic==2.10.4
python-multipart==0.0.18

